\name{panelSUR}
\alias{panelSUR}
\docType{package}
\title{panelSUR package: EC SUR system models estimation on (unbalanced) panel data}

\description{
Allows to estimate one-way and two-way error component SUR systems on unbalanced panel by GLS
    estimator with or without cross-equation restrictions.
}

\details{
\tabular{ll}{
Package: \tab panelSUR\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2024-03-03\cr
}
}

\references{
Biørn E (2004). Regression Systems for Unbalanced Panel Data: a Stepwise Maximum Likelihood Procedure. \emph{Journal of Econometrics}, \bold{122}(2), 181-291.

Platoni S, Barbieri L, Moro D, Sckokai P (2020). Heteroscedastic Stratified Two-way EC Models of Single
Equations and SUR Systems. \emph{Econometrics and Statistics}, \bold{15}, 46-66.

Platoni S, Sckokai P, Moro D, (2012a). “A Note on Two-way ECM Estimation of SUR Systems on Unbalanced Panel Data. \emph{Econometric Reviews}, \bold{31}(2), 119–141.
}
