\name{SURdata}
\docType{data}
\alias{SURdata}
\alias{data}
\title{Simulated data for a simultaneous equation system}
\description{
The \code{SURdata} dataset consists of an unbalanced panel comprising 100 individuals observed across four time periods for a total of 220 observations (n=100, T=4, N=220). In order to construct this unbalanced panel, the procedure currently used for rotating panels, in which there is approximately the same number of individuals every year, has been used: a fixed percentage of individuals (20\% in this case) is replaced each year, but they can re-enter the sample in the following years.
}

\usage{data(SURdata)}

\format{A large unbalanced panel dataset}
\source{Simulated data}

\keyword{datasets}
